переписаны view функции и сериализаторы

+++ backend/api/views.py
@@ -6,19 +6,25 @@ from djoser.serializers import UserSerializer
 from rest_framework.permissions import IsAuthenticated, SAFE_METHODS
 from rest_framework.decorators import action
 from rest_framework.response import Response
-from rest_framework import status, permissions, viewsets
+from rest_framework import status, permissions, viewsets,  exceptions
+from django.db.models import Sum
+
+
 from django_filters.rest_framework import DjangoFilterBackend
 from reportlab.pdfgen import canvas
 from collections import defaultdict
+
+from users.pagination import CustomPageNumberPagination
 
 from .filters import IngredientFilter
 from .permissions import IsAdminOrReadOnly, IsAuthorOrReadOnly
 from .serializers import (
     MyUserSerializer, MyUserCreateSerializer, 
-    UserFollowSerializer, FollowSerializer, TagSerializer, 
-    IngredientSerializer, RecipeSerializer, GetRecipeSerializer)
+    UserFollowSerializer, TagSerializer, 
+    IngredientSerializer, RecipeIngredientSerializer, CreateUpdateRecipeIngredientsSerializer,
+    GetRecipeSerializer, RecipeSerializer, ShortRecipeSerializer)
 from users.models import MyUser, Follow
-from recipes.models import Tag, Ingredient, Recipe, Favorite, ShopingList
+from recipes.models import Tag, Ingredient, Recipe, Favorite, ShopingList, RecipeIngredient
 
 
 User = get_user_model()
@@ -28,30 +34,11 @@ class MyUserViewSet(UserViewSet):
     queryset = MyUser.objects.all()
     serializer_class = MyUserSerializer
     permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
+    pagination_class = CustomPageNumberPagination
     
 
     @action(
-        methods=['GET', ],
-        detail=False,
-        url_path='me',
-        url_name='me',
-        permission_classes=[IsAuthenticated, ]
-    )
-    def me(self, request):
-        serializer = MyUserSerializer(
-            request.user, context={'request': request}
-        )
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-    def get_serializer_class(self):
-        if self.request.user.is_anonymous:
-            if self.request.method == 'POST':
-                return MyUserCreateSerializer
-            return UserSerializer
-        return super().get_serializer_class()
-
-    @action(
-        methods=['GET', ],
+        methods=['GET'],
         detail=False,
         url_path='subscriptions',
         url_name='subscriptions',
@@ -61,49 +48,63 @@ class MyUserViewSet(UserViewSet):
     def subscriptions(self, request):
         """Выдает авторов, на кого подписан пользователь"""
         user = request.user
-        following_authors = MyUser.objects.filter(following__user=user)
-        serializer = UserFollowSerializer(following_authors, many=True)
-        return Response(serializer.data)
+        queryset = user.following.values_list('author', flat=True).distinct()
+        pages = self.paginate_queryset(queryset)
+        serializer = UserFollowSerializer(
+            pages, many=True, context={'request': request}
+        )
+        return self.get_paginated_response(serializer.data)
 
     @action(
         methods=['POST', 'DELETE'],
-        detail=False,
+        detail=True,
         url_path='subscribe',
         url_name='subscribe',
         permission_classes=[IsAuthenticated, ])
 
-
-    def subscribe(self, request, id):
+    def subscribe(self, request, id=None):
         """Подписаться/отписаться на/от автора"""
-        user = request.user
-        author = get_object_or_404(User, id=id)
-        if request.method == 'POST':
-            serializer = FollowSerializer(
-                data={
-                    'user': user.id,
-                    'author': author.id
-                },
-                context={'request': request}
-            )
-            serializer.is_valid(raise_exception=True)
-            serializer.save()
-            return Response(
-                serializer.data,
-                status=status.HTTP_201_CREATED
-            )
-        subscription = get_object_or_404(
-            Follow,
-            user=user,
-            author=author
-        )
-        subscription.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)   
+        user = self.request.user
+        author = get_object_or_404(MyUser, pk=id)
 
+        if self.request.method == 'POST':
+            if user == author:
+                raise exceptions.ValidationError(
+                    'Подписка на самого себя запрещена.'
+                )
+            if Follow.objects.filter(
+                user=user,
+                author=author
+            ).exists():
+                raise exceptions.ValidationError('Подписка уже оформлена.')
 
-class TagViewSet(viewsets.ModelViewSet):
+            Follow.objects.create(user=user, author=author)
+            serializer = self.get_serializer(author)
+
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+        if self.request.method == 'DELETE':
+            if not Follow.objects.filter(
+                user=user,
+                author=author
+            ).exists():
+                raise exceptions.ValidationError(
+                    'Подписка не была оформлена, либо уже удалена.'
+                )
+            subscription = get_object_or_404(
+                Follow,
+                user=user,
+                author=author
+            )
+            subscription.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
+
+class TagViewSet(viewsets.ReadOnlyModelViewSet):
     """Viewset для объектов модели Tag"""
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
+    pagination_class = None
     permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
 
 
@@ -111,8 +112,6 @@ class IngredientViewSet(viewsets.ModelViewSet):
     """Viewset для объектов модели Ingredient"""
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerializer
-    permission_classes = (IsAdminOrReadOnly,)
-    filter_backends = [DjangoFilterBackend]
     filterset_class = IngredientFilter
     search_fields = ('^name', )
     pagination_class = None
@@ -122,20 +121,17 @@ class IngredientViewSet(viewsets.ModelViewSet):
 class RecipeViewSet(viewsets.ModelViewSet):
     """Viewset для объектов модели Recipe"""
     queryset = Recipe.objects.all()
-    serializer_class = RecipeSerializer
     permission_classes = (IsAuthorOrReadOnly, )
-    filter_backends = [DjangoFilterBackend]
+    filter_backends = (DjangoFilterBackend,)
     # filterset_class = RecipeFilter
+    pagination_class = CustomPageNumberPagination
 
     def get_serializer_class(self):
         """Определяет какой сериализатор использовать"""
-        if self.request.method in SAFE_METHODS and self.request.method == 'GET':
+        if self.action in ('create', 'partial_update'):
             return GetRecipeSerializer
-        return RecipeSerializer
 
-    def perform_create(self, serializer):
-        serializer.is_valid(raise_exception=True)
-        serializer.save(author=self.request.user)
+        return RecipeSerializer
 
 
     @action(
@@ -160,7 +156,7 @@ class RecipeViewSet(viewsets.ModelViewSet):
         url_name='recipes',
         permission_classes=[IsAuthenticated, ])
 
-    def recipe_delete(self, request):
+    def recipes_delete(self, request):
         recipe_id = request.query_params.get('id')
         if not recipe_id:
             return Response({'detail': 'Укажите идентификатор рецепта'}, status=status.HTTP_400_BAD_REQUEST)
@@ -174,27 +170,41 @@ class RecipeViewSet(viewsets.ModelViewSet):
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(
-        methods=['POST', 'DELETE'],
-        detail=False,
-        url_path='favorite',
-        url_name='favorite',
-        permission_classes=[IsAuthenticated, ])    
+            methods=['POST', 'DELETE'],
+            detail=True,
+            url_path='favorite',
+            url_name='favorite',
+            permission_classes=[permissions.IsAuthenticated]
+    )    
 
-
-    def favorite(self, request, id):
-        """Добавить/Удалить рецепт в/из избранное"""
-        try:
-            recipe = Recipe.objects.get(id=id)
-            favorite = Favorite.objects.get(user=request.user, recipe=recipe)
-            if request.method == 'DELETE':
-                favorite.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
-        except (Recipe.DoesNotExist, Favorite.DoesNotExist):
-            if request.method == 'POST':
-                Favorite.objects.create(user=request.user, recipe=recipe)
-                serializer = FollowSerializer(recipe)
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+    def favorite(self, request, pk=None):
+        """Для добавления/удаления в/из Избранное"""
+        user = self.request.user
+        recipe = get_object_or_404(Recipe, pk=pk)
+        if self.request.method == 'POST':
+            if Favorite.objects.filter(
+                user=user,
+                recipe=recipe
+            ).exists():
+                raise exceptions.ValidationError('Рецепт уже в избранном.')
+            Favorite.objects.create(user=user, recipe=recipe)
+            serializer = ShortRecipeSerializer(
+                recipe,
+                context={'request': request}
+            )
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if self.request.method == 'DELETE':
+            if not Favorite.objects.filter(
+                user=user,
+                recipe=recipe
+            ).exists():
+                raise exceptions.ValidationError(
+                    'Рецепта нет в избранном, либо он уже удален.'
+                )
+            favorite = get_object_or_404(Favorite, user=user, recipe=recipe)
+            favorite.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
 
     @action(
@@ -207,63 +217,69 @@ class RecipeViewSet(viewsets.ModelViewSet):
 
     def download_shopping_cart(self, request):
         """Скачать список покупок"""
-        # Получаем список всех рецептов, сохраненных в списке покупок текущего пользователя
-        shop_list = ShopingList.objects.filter(user=request.user)
-        recipes = [item.recipe for item in shop_list]
+        shopping_cart = ShopingList.objects.filter(user=self.request.user)
+        recipes = [item.recipe.id for item in shopping_cart]
+        buy_list = RecipeIngredient.objects.filter(
+            recipe__in=recipes
+        ).values(
+            'ingredient'
+        ).annotate(
+            amount=Sum('amount')
+        )
+        buy_list_text = 'Список покупок:\n\n'
+        for item in buy_list:
+            ingredient = Ingredient.objects.get(pk=item['ingredient'])
+            amount = item['amount']
+            buy_list_text += (
+                f'{ingredient.name}, {amount} '
+                f'{ingredient.measurement_unit}\n'
+            )
+        response = HttpResponse(buy_list_text, content_type="text/plain")
+        response['Content-Disposition'] = (
+            'attachment; filename=shopping-list.txt'
+        )
 
-        # Создаем словарь, где ключами будут имена ингредиентов вместе с их единицами измерения, а значениями будут суммарные количество ингредиентов, необходимых для приготовления всех рецептов
-        ingredients_list = defaultdict(float)
-        for recipe in recipes:
-            for recipe_ingredient in recipe.recipe_ingredients.all():
-                ingredient = recipe_ingredient.ingredient
-                key = f"{ingredient.name} - {ingredient.measurement_unit}"
-                ingredients_list[key] += recipe_ingredient.amount
-
-        # Формируем текстовое представление полученного словаря
-        shopping_cart = ""
-        for key, value in ingredients_list.items():
-            shopping_cart += f"{key} - {value}\n"
-
-    # Создаем PDF-файл, если нужно
-        if request.GET.get('pdf'):
-            pdf_file = io.BytesIO()
-            p = canvas.Canvas(pdf_file)
-            p.drawString(100, 750, "Shopping List")
-            p.drawString(100, 700, shopping_cart)
-            p.save()
-
-            # Отправляем PDF-файл пользователю
-            response = HttpResponse(pdf_file.getvalue(), content_type='application/pdf')
-            response['Content-Disposition'] = 'attachment; filename="shopping_list.pdf"'
-            return response
-
-        # Отправляем текстовый файл пользователю
-        response = HttpResponse(shopping_cart, content_type='text/plain')
-        response['Content-Disposition'] = 'attachment; filename="shopping_list.txt"'
         return response
 
 
     @action(
         methods=['POST', 'DELETE'],
-        detail=False,
+        detail=True,
         url_path='shopping_cart',
         url_name='shopping_cart',
         permission_classes=[IsAuthenticated, ])    
 
-    def shopping_cart(self, request, id ):
+    def shopping_cart(self, request, pk=None ):
         """Добавить / удалить рецепт в список покупок"""
-        try:
-            recipe = Recipe.objects.get(id=id)
-            shop_cart = ShopingList.objects.get(user=request.user, recipe=recipe)
-            if request.method == 'DELETE':
-                shop_cart.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
-        except (Recipe.DoesNotExist, ShopingList.DoesNotExist):
-            if request.method == 'POST':
-                shop_cart = ShopingList.objects.create(user=request.user, recipe=recipe)
-                serializer = FollowSerializer(shop_cart.recipe)
-                return Response({'message': 'Рецепт успешно добавлен в список покупок!', 'data': serializer.data}, status=status.HTTP_201_CREATED)
-            return Response({'message': 'Ошибка! Рецепт не найден в списке покупок', 'data': {}}, status=status.HTTP_404_NOT_FOUND)
-
-        return Response({'message': 'Bad request'}, status=status.HTTP_400_BAD_REQUEST)
-
+        user = self.request.user
+        recipe = get_object_or_404(Recipe, pk=pk)
+        if self.request.method == 'POST':
+            if ShopingList.objects.filter(
+                user=user,
+                recipe=recipe
+            ).exists():
+                raise exceptions.ValidationError(
+                    'Рецепт уже в списке покупок.'
+                )
+            ShopingList.objects.create(user=user, recipe=recipe)
+            serializer = ShortRecipeSerializer(
+                recipe,
+                context={'request': request}
+            )
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if self.request.method == 'DELETE':
+            if not ShopingList.objects.filter(
+                user=user,
+                recipe=recipe
+            ).exists():
+                raise exceptions.ValidationError(
+                    'Рецепта нет в списке покупок, либо он уже удален.'
+                )
+            shopping_cart = get_object_or_404(
+                ShopingList,
+                user=user,
+                recipe=recipe
+            )
+            shopping_cart.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
